fix check existance of collection

+++ initDb.js
@@ -11,7 +11,9 @@ async function init (){
 	try {
 		client = await MongoClient.connect(url);
 		const db = client.db(dbName);
-		await  db.collection(collectionName).drop();
+		if (db.collections.length > 0){
+			await  db.collection(collectionName).drop();
+		}
 		const collection = db.collection(collectionName);
 		let jsonBeers = JSON.parse(jsonData.toString());
